third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-reporting-for-SharedWorker-ServiceWorkerContainer-register.https-expected.txt

This is a testharness.js-based test.
[FAIL] No violation reported for ServiceWorkerContainer register with TrustedScriptURL.
  assert_equals: no exception thrown expected null but got object "TypeError: Cannot read properties of undefined (reading 'getRegistration')"
[FAIL] Violation report for ServiceWorkerContainer register with plain string.
  assert_equals: a single violation reported expected 1 but got 0
Harness: the test ran to completion.
